feat: create projects page
