+++ src/components/Landing/ProjectSection/project-section.module.css
@@ -17,122 +17,3 @@
     max-width: 500px;
     margin-top: 0;
 }
-
-.project {
-    display: flex;
-    margin: 10rem 0;
-}
-
-.projectRight {
-    flex-flow: row-reverse;
-    justify-content: space-between;
-}
-
-.projectTable {
-    border-collapse: collapse;
-    width: 100%;
-    margin-bottom: 2rem;
-}
-  
-.projectTable td, .projectTable th {
-border: 1px solid #ddd;
-padding: 6px;
-}
-
-.projectTable tr:nth-child(even){background-color: #f2f2f2;}
-
-.projectTable tr:hover {background-color: #ddd;}
-
-.textContainer {
-    margin: 0 5rem;
-    height: 100%;
-    display: flex;
-}
-
-.buttonContainer {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    margin-right: 8rem;
-}
-
-.buttonContainer p {
-    margin-top: 250px;
-    width: 100px;
-    font-size: small;
-}
-
-.textContentContainer {
-    max-width: 500px;
-}
-
-.textContentContainer h2 {
-    color: #444;
-    font-size: var(--text-3xl);
-    margin-bottom: 2rem;
-}
-
-.gallery {
-    width: 700px;
-}
-
-@media (max-width: 1024px) {
-    .buttonContainer {
-        margin-right: 3rem;
-    }
-    .buttonContainer p {
-        margin-top: 150px;
-        width: 100px;
-        font-size: smaller;
-    }
-    .textContentContainer h2 {
-        font-size: var(--text-2xl);
-    }
-    .textContentContainer h4 {
-        font-size: var(--text-2xl);
-        margin-bottom: 1rem;
-    }
-
-    .textContentContainer p {
-        font-size: smaller;
-    }
-}
-
-@media (max-width: 768px) {
-    .gallery {
-        width: 100%;
-        margin-bottom: 3rem;
-    }
-    .project {
-        flex-wrap: wrap;
-    }
-    .buttonContainer {
-        margin-right: none;
-    }
-    .buttonContainer p {
-        display: none;
-    }
-    .projectTable {
-        width: 100%;
-        margin-bottom: 0rem;
-    }
-    .textContainer {
-        margin: 0;
-    }
-    .textContentContainer {
-        max-width: 300px;
-    }
-    .textContainer {
-        flex-wrap: wrap;
-        margin: 0 1rem;
-    }
-    .buttonContainer {
-        margin-right: 0rem;
-    }
-    .root {
-        padding-bottom: 3rem;
-    }
-    .buttonContainer p {
-        display: none;
-    }
-}
